Move fitting pieces from dynspec and conjspec to modeling
